11h00-23/1

+++ src/app/timeline/timeline.component.ts
@@ -118,7 +118,7 @@ export class TimelineComponent implements OnInit {
     this.statusService.addReplyStatus(statusId, this.auth.currentUserValue.userName,  this.createReplyStatus()).subscribe((data) => {
       console.log('OK');
       this.replyStatusForm.reset();
-      this.getStatuses(userName);
+      this.getStatuses(this.userNamePath);
     });
   }
 
@@ -135,10 +135,14 @@ export class TimelineComponent implements OnInit {
 
   // @ts-ignore
   deleteStatus(statusId) {
-    // @ts-ignore
-    this.statusService.deleteStatus(statusId).subscribe(data => {
-      console.log('delete', data);
-    });
+    if (this.userName === this.auth.currentUserValue.userName) {
+      // @ts-ignore
+      this.statusService.deleteStatus(statusId).subscribe(data => {
+        console.log('delete', data);
+
+        this.getStatuses(this.userName);
+      });
+    }
   }
 
   // @ts-ignore
@@ -164,7 +168,6 @@ export class TimelineComponent implements OnInit {
     this.router.navigate(['status-form', this.userName]);
     console.log('OK');
   }
-
 
   likeStatus(statusId: number, userName: string){
     this.likeService.likeStatus(statusId, this.auth.currentUserValue.userName ).subscribe(data => {
